requirements.txt
fastapi
langchain
langchain-core
langchain-openai
langchain-community
langchain-huggingface
pypdf
chroma
python-dotenv
chromadb
pydantic    
uvicorn
sentence-transformers